+++ src/UniAlteri/States/bootstrap_composer.php
@@ -26,12 +26,15 @@ use UniAlteri\States\Factory\FactoryInterface;
 
 defined('UA_STATES_PATH') || define('UA_STATES_PATH', __DIR__);
 
-//Needed for test, but we can use your own autoloader to load file of this lib.
-//This lib respects PSR-0, PSR-1 and PSR-2
-$iniFile = dirname(dirname(dirname(__DIR__))).DIRECTORY_SEPARATOR.'autoloader_psr0.php';
-if (is_readable($iniFile)) {
-    include_once $iniFile;
+//Shortcut for DIRECTORY_SEPARATOR
+defined('DS') || define('DS', DIRECTORY_SEPARATOR);
+
+//Use composer has default auto loader
+$composerFile = __DIR__.'/../../../../vendor/autoload.php';
+if (!file_exists($composerFile)) {
+    $composerFile = __DIR__.'/../../../vendor/autoload.php';
 }
+$composerInstance = require $composerFile;
 
 //Initial DI Container
 $diContainer = new DI\Container();
@@ -42,17 +45,17 @@ $diContainer->registerInstance(FactoryInterface::DI_FACTORY_REPOSITORY, new DI\C
 /*
  * Service to generate a finder for Stated class factory
  * @param DI\ContainerInterface $container
- * @return Loader\FinderIntegrated
+ * @return Loader\FinderComposerIntegrated
  * @throws Exception\UnavailableFactory if the local factory is not available
  */
-$finderService = function (DI\ContainerInterface $container) {
+$finderService = function (DI\ContainerInterface $container) use ($composerInstance) {
     if (false === $container->testEntry(Factory\FactoryInterface::DI_FACTORY_NAME)) {
         throw new Exception\UnavailableFactory('Error, the factory is not available into container');
     }
 
     $factory = $container->get(Factory\FactoryInterface::DI_FACTORY_NAME);
 
-    return new Loader\FinderIntegrated($factory->getStatedClassName(), $factory->getPath());
+    return new Loader\FinderComposerIntegrated($factory->getStatedClassName(), $factory->getPath(), $composerInstance);
 };
 
 //Register finder generator
@@ -70,8 +73,7 @@ $injectionClosureService = function () {
 $diContainer->registerService(States\StateInterface::INJECTION_CLOSURE_SERVICE_IDENTIFIER, $injectionClosureService);
 
 //Stated class loader, initialize
-$includePathManager = new Loader\IncludePathManager();
-$loader = new Loader\LoaderStandard($includePathManager);
+$loader = new Loader\LoaderComposer($composerInstance);
 $loader->setDIContainer($diContainer);
 
 //Register loader into container